Reverts change from current to wanted
Current is the one specified in the package.json and for the intents of this utility I feel this is the better choice because the goal is to make you aware and more proactive in updating and researching.

Using wanted would be good to compare the actual installed version versus latest but I feel that this isn't as proactive. This could be something supported later by config.