Updates dependencies
There was a minor break now that TypeScript
assesses the error in a catch as `unknown`.

Also converted the child process `execFileSync` to the async version. The callback of the async version is already typed to the correct node error saving me from writing type guards to come to the same conclusion using the synchronous version.